+++ aztk/models/plugins/common/hdfs/hdfs.sh
