+++ aztk/models/plugins/common/jupyter/__init__.py
@@ -1,0 +1,1 @@
+from .configuration import *
