+++ aztk/models/plugins/common/jupyter/jupyter.sh
