--- aztk/models/plugins/common/jupyter_plugin.py
@@ -1,7 +1,0 @@
-from aztk.models.plugins.plugin_configuration import PluginConfiguration
-
-class JupyterPlugin(PluginConfiguration):
-    def __init__(self):
-        super().__init__(
-            name="jupyter",
-        )
