+++ aztk/models/plugins/common/rstudio_server/__init__.py
@@ -1,0 +1,1 @@
+from .configuration import *
