+++ aztk/models/plugins/common/rstudio_server/configuration.py
@@ -1,0 +1,25 @@
+import os
+from aztk.models.plugins.plugin_configuration import PluginConfiguration
+from aztk.models.plugins.plugin_file import PluginFile
+from aztk.utils import constants
+
+dir_path = os.path.dirname(os.path.realpath(__file__))
+
+
+class RStudioServerPlugin(PluginConfiguration):
+    def __init__(self, version="1.1.383"):
+        super().__init__(
+            name="rstudio_server",
+            ports=[
+                PluginPort(
+                    internal=8787,
+                    public=True,
+                ),
+            ],
+            run_on=PluginRunTarget.Master,
+            execute="rstudio_server.sh",
+            files=[
+                PluginFile.from_local(os.path.join(dir_path, "rstudio_server.sh"), "rstudio_server.sh"),
+            ],
+            env=dict(RSTUDIO_SERVER_VERSION=version),
+        )
