+++ aztk/models/plugins/common/rstudio_server/rstudio_server.sh
