--- base_plugins/jupyter/main.py
@@ -1,18 +1,0 @@
-from aztk.models.plugins import PluginDefinition, PluginPort, PluginRunTarget
-
-
-def definition():
-    return PluginDefinition(
-        name="jupyter",
-        ports=[
-            PluginPort(
-                internal=8888,
-                public=True,
-            ),
-        ],
-        run_on=PluginRunTarget.All,
-        execute="jupyter.sh",
-        files=[
-            "jupyter.sh",
-        ],
-    )
